pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = "Habit Tracker - помогает приобретению новых полезных привычек и искоренению старых плохих привычек."
authors = ["Maxim Ipatov <ipatovmaksim.1978@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
Django = "^4.2.17"
django-bootstrap5 = "^24.3"
djlint = "^1.36.3"
pillow = "^11.0.0"
faker = "^33.1.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.10"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
stripe = "^11.4.1"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

max-complexity = 10
per-file-ignores = { "__init__.py" = "F401" }

[tool.flake8]
line-length = 79
extend-ignore = ["E203"]
exclude = [
    "/__pycache__/",
    "**/migrations/",
    "env/",
    "venv/",
    ".env/",
    ".venv/",
    "tests",
    ".idea",
    ".vscode",
    ".git",
]

[tool.black]
line-length = 79
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 79
force_grid_wrap = 0
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true

